Moving alertdetaillogs.js
